performance: Change placement conditions system
This change is pretty drastic, but yields very sizable improvements.

Before, we iterated over every block in a blueprint and checked each
block against every condition. This had a lot of overhead in cases where
conditions only cared about very small sections of data and prevented
us from ordering the conditions properly to avoid unnecessary overhead.

This commit changes PlacementCondition to instead take a baked blueprint,
the offset into the world, and the world itself. PlacementConditions are
responsible for iterating and checking their own data, which allows for
a lot of optimizations to take place and for us to skip iterating
unnecessary data, being our ultimate performance killer.