performance: Implement WorldSlice class, optimize block filtering
WorldSlice provides a faster means of fetching and updating blocks from
the world during the population phase by avoiding expensive operations in
the World class and by directly storing the ExtendedBlockStorage segments
in a flat array. For generators that can take advantage of it, there are
also extremely optimized paths for replacing blocks and performing light
checks/updates only when we know it's necessary. This provides a modest
boost to ore generation as a whole.

Added a "touched" flag to ChunkSegmentMask so we can avoid iterating
over empty ChunkSegmentMasks when copy blocks into ChunkDataContainer.

Changed IPositionRecords to not provide facilities to filter matching
blocks. This creates significant overhead when multiple data entries
exist and causes excessive iteration.

Additionally, ICreationData#shouldCreate no longer has access to
the block state trying to be placed, as it's very wasteful to work down
the chain only to terminate at the last moment. Instead, the check is
performed for each position as soon as possible to improve performance.

As a last note, this also changes BlockUtil and various predicates to
perform less checks against blocks as there was a lot of duplication on
things such as material and bounds checks.