performance: Separate block IDs from block metadata in BlockDataContainer
Use the Block class as the hash key in our lookup functions

This is generally much faster than using IBlockState as a hash key because
IBlockState includes all the properties and other data in its hashcode,
which we don't care about.

Additionally, avoid an unnecessary additional lookup in setBlock and
again when copying blocks.